+++ internal/framework/schema_common.go
@@ -1,3 +1,5 @@
+// Copyright © 2025 Ping Identity Corporation
+
 package framework
 
 import (
@@ -169,6 +171,23 @@ func Attr_DataSourceReturnIDsByElement(description SchemaAttributeDescription, e
 	}
 
 	return schema.ListAttribute{
+		Description:         description.Description,
+		MarkdownDescription: description.MarkdownDescription,
+		Computed:            true,
+		ElementType:         elementType,
+	}
+}
+
+func Attr_DataSourceReturnIDsSet(description SchemaAttributeDescription) schema.SetAttribute {
+	return Attr_DataSourceReturnIDsByElementSet(description, pingonetypes.ResourceIDType{})
+}
+
+func Attr_DataSourceReturnIDsByElementSet(description SchemaAttributeDescription, elementType attr.Type) schema.SetAttribute {
+	if description.MarkdownDescription == "" {
+		description.MarkdownDescription = description.Description
+	}
+
+	return schema.SetAttribute{
 		Description:         description.Description,
 		MarkdownDescription: description.MarkdownDescription,
 		Computed:            true,