+++ internal/service/sso/data_source_populations.go
@@ -1,3 +1,5 @@
+// Copyright © 2025 Ping Identity Corporation
+
 package sso
 
 import (
@@ -162,7 +164,7 @@ func (r *PopulationsDataSource) Read(ctx context.Context, req datasource.ReadReq
 	} else {
 		resp.Diagnostics.AddError(
 			"Missing parameter",
-			"Cannot find the requested population. population_id or name must be set.",
+			"Cannot filter the populations. scim_filter or data_filters must be set.",
 		)
 		return
 	}